+++ Buscador.html
@@ -82,18 +82,19 @@
   <div class="Tarjeta" id="crearTarjeta"> </div>
       
    
-     
-
-    <!-- JS -->
-    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-    <script src="js/main buscador.js" type="module"></script>
-    
-    
-     <!-- JS Librerias-->
+      <!-- JS Librerias-->
   <script src="lib/jquery/jquery.min.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.4.1/dist/js/bootstrap.min.js"></script>
   <script src="lib/counterup/jquery.counterup.js"></script>
   <script src="lib/typed/typed.min.js"></script>
-  <script src="js/main index.js"></script>
+
+    <!-- JS -->
+    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+    <script src="js/mainBuscador.js" type="module"></script>
+    <script src="cards.js" type="module"></script>
+    <script src="js/mainIndex.js"></script>
+    
+    
+  
 </body>
 </html>
