+++ index.html
@@ -77,7 +77,7 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.4.1/dist/js/bootstrap.min.js"></script>
   <script src="lib/counterup/jquery.counterup.js"></script>
   <script src="lib/typed/typed.min.js"></script>
-  <script src="js/main index.js"></script>
+  <script src="js/mainIndex.js"></script>
 
 </body>
 </html>
