+++ package.json
@@ -2,7 +2,6 @@
   "name": "animapi",
   "version": "0.1.0",
   "description": "consumo de api con grafico e imagenes",
-  "type": "module",
   "main": "index.js",
   "directories": {
     "lib": "lib"
